Added functionality only authenticated users, or users with cookie to be able to edit or delete comments. Not authenticated users can edit only their last comment.

+++ polls_app/core/views_mixins.py
@@ -1,23 +1,29 @@
 from rest_framework import status
 from rest_framework.response import Response
 
-from polls_app.core.services import get_object_and_check_permission_service
+from polls_app.core.services import check_comment_ownership_service
 
 
 class UpdateDeleteMixin:
 
     def patch(self, request, *args, **kwargs):
         """
-        PATCH request edit an object with the given id.
+        PATCH request to edit an object with the given ID.
         """
         instance = self.get_object()
+
+        # Check ownership for authenticated and anonymous users
+        if not check_comment_ownership_service(request, instance):
+            return Response(
+                {"detail": "You do not have permission to edit this comment."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         serializer = self.get_serializer(instance, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
         serializer.save()
 
         if getattr(instance, '_prefetched_objects_cache', None):
-            # If 'prefetch_related' has been applied to a queryset, we need to
-            # forcibly invalidate the prefetch cache on the instance.
             instance._prefetched_objects_cache = {}
 
         return Response(
@@ -27,13 +33,23 @@ class UpdateDeleteMixin:
 
     def delete(self, request, *args, **kwargs):
         """
-        DELETE request delete and object with the given id.
+        DELETE request to delete an object with the given ID.
         """
         instance = self.get_object()
+
+        # Check ownership for authenticated and anonymous users
+        if not check_comment_ownership_service(request, instance):
+            return Response(
+                {"detail": "You do not have permission to delete this comment."},
+                status=status.HTTP_403_FORBIDDEN,
+            )
+
         instance.delete()
 
         return Response(
             {"message": "Successfully deleted"},
             status=status.HTTP_204_NO_CONTENT,
         )
+
+
